nuking the repo
