+++ components/Hamburger.js
@@ -1,0 +1,13 @@
+import { IconButton } from "@chakra-ui/react";
+import { AiOutlineEllipsis } from "react-icons/ai";
+
+export default function Hamburger() {
+
+  return (
+    <IconButton
+      aria-label="Menu"
+      variant="ghost"
+      icon={<AiOutlineEllipsis />}
+    />
+  );
+}
