+++ context/AppContext.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+const AppContext = React.createContext();
+
+export default AppContext;
