+++ next.config.js
@@ -1,3 +1,9 @@
 module.exports = {
-  reactStrictMode: true,
-}
+
+  webpack: function (config, options) {
+    config.experiments = {
+      topLevelAwait: true
+    };
+    return config;
+  },
+};
