+++ package.json
@@ -11,6 +11,7 @@
     "@chakra-ui/react": "^1.7.2",
     "@emotion/react": "^11.6.0",
     "@emotion/styled": "11.6.0",
+    "@walletconnect/web3-provider": "^1.6.6",
     "formik": "^2.2.9",
     "framer-motion": "^4.1.17",
     "install": "^0.13.0",
@@ -20,6 +21,7 @@
     "react-dom": "17.0.2",
     "react-icons": "^4.3.1",
     "web3": "^1.6.1",
+    "web3modal": "^1.9.4",
     "yup": "^0.32.11"
   },
   "devDependencies": {
