+++ utils/addresses.js
@@ -1,0 +1,4 @@
+// factory addresses
+export const factory_arbitrum = "0xd53B46aE3781904F1f61CF38Fd9d4F47A7e9242B";
+export const factory_polygon = "0x582eAF6a83E55d60615A5FfB80913bE5c1724c41";
+export const factory_rinkeby = "0x4941c423ad4E297932C5E32aB40894544292961F";
