+++ utils/proposalTypes.js
@@ -1,0 +1,1 @@
+["MINT","BURN","CALL","PERIOD","QUORUM","SUPERMAJORITY","PAUSE","EXTENSION"]
